Half Of Yurok Village Rebuilt
Replaced rock pieces with new treetops, set walkway pieces around the water, and built new house BPs to place around the levels. Also fixed the grid for the left half of the map with Henry and adjusted collision boxes.

Currently, the lighting is broken and that may be the result of a skybox being removed.