Yurok Village BPs
Yurok Village now has homes you can enter. There is a "long house" which is the shape of an L, a small house that is just a square, and the Canopy Inn is complete with two floors.

Separately, the bridge pieces have been placed which means no more white blocks. Additional support trees placed beneath the bridge as well. Another small change I did was adjust the angles of some of the fishing poles so they're not fully erect.